+++ CHANGELOG.md
@@ -15,3 +15,4 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Changed
 - Use new nrfutil version
+- Update fota_sender_with_driver example to use new mira_fota_set_driver API